--- src/libraries/Common/src/Interop/Windows/Crypt32/Interop.CERT_CHAIN_PARA.cs
@@ -1,45 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System;
-using System.Runtime.InteropServices;
-
-internal static partial class Interop
-{
-    internal static partial class Crypt32
-    {
-        [StructLayout(LayoutKind.Sequential)]
-        internal unsafe struct CERT_CHAIN_PARA
-        {
-            public int cbSize;
-            public CERT_USAGE_MATCH RequestedUsage;
-            public CERT_USAGE_MATCH RequestedIssuancePolicy;
-            public int dwUrlRetrievalTimeout;
-            public int fCheckRevocationFreshnessTime;
-            public int dwRevocationFreshnessTime;
-            public FILETIME* pftCacheResync;
-            public int pStrongSignPara;
-            public int dwStrongSignFlags;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct CERT_USAGE_MATCH
-        {
-            public CertUsageMatchType dwType;
-            public CTL_USAGE Usage;
-        }
-
-        internal enum CertUsageMatchType : int
-        {
-            USAGE_MATCH_TYPE_AND = 0x00000000,
-            USAGE_MATCH_TYPE_OR  = 0x00000001,
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct CTL_USAGE
-        {
-            public int cUsageIdentifier;
-            public IntPtr rgpszUsageIdentifier;
-        }
-    }
-}
