+++ src/libraries/Common/src/Interop/Windows/Crypt32/Interop.CertGetIntendedKeyUsage.cs
@@ -10,7 +10,8 @@ internal static partial class Interop
         // Note: It's somewhat unusual to use an API enum as a parameter type to a P/Invoke but in this case, X509KeyUsageFlags was intentionally designed as bit-wise
         // identical to the wincrypt CERT_*_USAGE values.
         [GeneratedDllImport(Libraries.Crypt32, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal static unsafe partial bool CertGetIntendedKeyUsage(CertEncodingType dwCertEncodingType,
+        internal static unsafe partial bool CertGetIntendedKeyUsage(
+            CertEncodingType dwCertEncodingType,
             CERT_INFO* pCertInfo,
             out X509KeyUsageFlags pbKeyUsage,
             int cbKeyUsage);
