+++ src/libraries/Common/src/Interop/Windows/Crypt32/Interop.CryptDecodeObjectPointer_IntPtr.cs
@@ -9,7 +9,8 @@ internal static partial class Interop
     internal static partial class Crypt32
     {
         [GeneratedDllImport(Libraries.Crypt32, EntryPoint = "CryptDecodeObject", CharSet = CharSet.Unicode, SetLastError = true)]
-        internal static unsafe partial bool CryptDecodeObjectPointer(CertEncodingType dwCertEncodingType,
+        internal static unsafe partial bool CryptDecodeObjectPointer(
+            CertEncodingType dwCertEncodingType,
             IntPtr lpszStructType,
             byte[] pbEncoded,
             int cbEncoded,
