+++ src/libraries/Common/src/Interop/Windows/Crypt32/Interop.CryptImportPublicKeyInfoEx2.cs
@@ -9,7 +9,8 @@ internal static partial class Interop
     internal static partial class Crypt32
     {
         [GeneratedDllImport(Libraries.Crypt32, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal static unsafe partial bool CryptImportPublicKeyInfoEx2(CertEncodingType dwCertEncodingType,
+        internal static unsafe partial bool CryptImportPublicKeyInfoEx2(
+            CertEncodingType dwCertEncodingType,
             CERT_PUBLIC_KEY_INFO* pInfo,
             CryptImportPublicKeyInfoFlags dwFlags,
             void* pvAuxInfo,
